+++ CHANGELOG.rst
@@ -36,3 +36,4 @@ Added
 - Non-Max suppression (NMS) function added.
 - Mean average precision calculation added.
 - Loss function returns a dictionary of all losses.
+- Added Scheduler class for flexible use of schedulers.